+++ queries/initialize_tables.sql
@@ -52,22 +52,27 @@ VALUES ('AK', 'Alaska'),
        ('WA', 'Washington'),
        ('WV', 'West Virginia'),
        ('WI', 'Wisconsin'),
-       ('WY', 'Wyoming');
+       ('WY', 'Wyoming')
+ON CONFLICT DO NOTHING;
 
 
 -- Titles
 INSERT INTO pdl.titles(name, description)
 VALUES ('Title I: Commodities',
-        'Title I, Commodities cover price and income support for the farmers who raise widely-produced and traded non-perishable crops, like corn, soybeans, wheat, cotton and rice – as well as dairy and sugar. The title also includes agricultural disaster assistance. The map shows the total benefits paid to farmers from of the commodities programs by state from 2018-2022.');
+        'Title I, Commodities cover price and income support for the farmers who raise widely-produced and traded non-perishable crops, like corn, soybeans, wheat, cotton and rice – as well as dairy and sugar. The title also includes agricultural disaster assistance. The map shows the total benefits paid to farmers from of the commodities programs by state from 2018-2022.')
+ON CONFLICT DO NOTHING;
 INSERT INTO pdl.titles(name, description)
 VALUES ('Title II: Conservation',
-        'Title II of the law authorizes the Farm Bill’s conservation programs. The programs in this title programs help agricultural producers and landowners adopt conservation activities on private farm and forest lands. In general, conservation activities are intended to protect and improve water quality and quantity, soil health, wildlife habitat, and air quality. The map shows the total benefit of the conservation program by state from 2018-2022.');
+        'Title II of the law authorizes the Farm Bill’s conservation programs. The programs in this title programs help agricultural producers and landowners adopt conservation activities on private farm and forest lands. In general, conservation activities are intended to protect and improve water quality and quantity, soil health, wildlife habitat, and air quality. The map shows the total benefit of the conservation program by state from 2018-2022.')
+ON CONFLICT DO NOTHING;
 INSERT INTO pdl.titles(name, description)
 VALUES ('Title IV: Nutrition',
-        'The Supplemental Nutrition Assistance Program [SNAP] provides financial assistance to low-income families to help cover the cost of food. Benefits can only be used to purchase food products and are provided in electronic format similar to a credit card and known as the Electronic Benefit Transfer (EBT) card. The map shows the total SNAP costs of the nutrition title by state from 2018-2022.');
+        'The Supplemental Nutrition Assistance Program [SNAP] provides financial assistance to low-income families to help cover the cost of food. Benefits can only be used to purchase food products and are provided in electronic format similar to a credit card and known as the Electronic Benefit Transfer (EBT) card. The map shows the total SNAP costs of the nutrition title by state from 2018-2022.')
+ON CONFLICT DO NOTHING;
 INSERT INTO pdl.titles(name, description)
 VALUES ('Title IX: Crop Insurance',
-        'Crop Insurance provides farmers with the option to purchase insurance policies on the acres of crops they plant to help manage the risks of farming, including to indemnify against losses in yields, crop or whole farm revenue, crop margins and other risks. The program also offsets the cost of the insurance policies through premium subsidies. In addition, the program provides Administrative and Operating (A&O) subsidies to the private crop insurance companies who provide federal crop insurance to farmers. The map shows the total farmer net benefitbenefit of the crop insurance program by state from 2018-2022.');
+        'Crop Insurance provides farmers with the option to purchase insurance policies on the acres of crops they plant to help manage the risks of farming, including to indemnify against losses in yields, crop or whole farm revenue, crop margins and other risks. The program also offsets the cost of the insurance policies through premium subsidies. In addition, the program provides Administrative and Operating (A&O) subsidies to the private crop insurance companies who provide federal crop insurance to farmers. The map shows the total farmer net benefit of the crop insurance program by state from 2018-2022.')
+ON CONFLICT DO NOTHING;
 
 -- Title: 1 - Subtitles
 INSERT INTO pdl.subtitles(title_id, name)
@@ -143,6 +148,82 @@ INSERT INTO pdl.sub_programs(program_id, name)
 SELECT id, 'Agriculture Risk Coverage Individual Coverage (ARC-IC)'
 FROM pdl.programs
 WHERE name = 'Agriculture Risk Coverage (ARC)';
+
+-- Title: 2 - Programs
+INSERT INTO pdl.programs(title_id, name)
+SELECT id, 'Environmental Quality Incentives Program (EQIP)'
+FROM pdl.titles
+WHERE name = 'Title II: Conservation';
+
+INSERT INTO pdl.programs(title_id, name)
+SELECT id, 'Conservation Stewardship Program (CSP)'
+FROM pdl.titles
+WHERE name = 'Title II: Conservation';
+
+INSERT INTO pdl.programs(title_id, name)
+SELECT id, 'Conservation Reserve Program (CRP)'
+FROM pdl.titles
+WHERE name = 'Title II: Conservation';
+
+INSERT INTO pdl.programs(title_id, name)
+SELECT id, 'Agricultural Conservation Easement Program (ACEP)'
+FROM pdl.titles
+WHERE name = 'Title II: Conservation';
+
+INSERT INTO pdl.programs(title_id, name)
+SELECT id, 'Regional Conservation Partnership Program (RCPP)'
+FROM pdl.titles
+WHERE name = 'Title II: Conservation';
+
+-- EQIP Practice Categories
+WITH eqip_program_id AS (SELECT id
+                         FROM pdl.programs
+                         WHERE name = 'Environmental Quality Incentives Program (EQIP)')
+INSERT
+INTO pdl.practice_categories(name, display_name, category_grouping, program_id, is_statutory_category)
+VALUES ('Land Management', 'Land management', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Forest Management', 'Forest management', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Structural', 'Structural', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Soil Remediation', 'Soil remediation', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Vegetative', 'Vegetative', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Other Improvement', 'Other improvement', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Soil Testing', 'Soil testing', '(6)(A) Practices', (SELECT id from eqip_program_id), TRUE),
+
+       ('Conservation Planning Assessment', 'Conservation planning assessment', '(6)(B) Practices',
+        (SELECT id from eqip_program_id), TRUE),
+       ('Other Planning', 'Other planning', '(6)(B) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Resource-conserving Crop Rotation', 'Resource-conserving crop rotation', '(6)(B) Practices',
+        (SELECT id from eqip_program_id), TRUE),
+       ('Soil Health', 'Soil health', '(6)(B) Practices', (SELECT id from eqip_program_id), TRUE),
+       ('Comprehensive Nutrient Management', 'Comprehensive Nutrient Mgt.', '(6)(B) Practices',
+        (SELECT id from eqip_program_id), TRUE);
+
+
+-- CSP Practice Categories
+WITH csp_program_id AS (SELECT id
+                        FROM pdl.programs
+                        WHERE name = 'Conservation Stewardship Program (CSP)')
+INSERT
+INTO pdl.practice_categories(name, display_name, category_grouping, program_id, is_statutory_category)
+VALUES ('Cropland', 'Cropland', '2014 Eligible Land', (SELECT id from csp_program_id), FALSE),
+       ('Grassland', 'Grassland', '2014 Eligible Land', (SELECT id from csp_program_id), FALSE),
+       ('Rangeland', 'Rangeland', '2014 Eligible Land', (SELECT id from csp_program_id), FALSE),
+       ('Pastureland', 'Pastureland', '2014 Eligible Land', (SELECT id from csp_program_id), FALSE),
+       ('Non-industrial Private Forestland', 'Non-industrial private forestland', '2014 Eligible Land',
+        (SELECT id from csp_program_id), FALSE),
+       ('Other: Supplemental, Adjustment & Other', 'Other: supplemental, adjustment & other', '2014 Eligible Land',
+        (SELECT id from csp_program_id), FALSE),
+       ('Structural', 'Structural', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Vegetative', 'Vegetative', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Land Management', 'Land management', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Forest Management', 'Forest management', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Soil Remediation', 'Soil remediation', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Existing Activity Payments', 'Existing activity payments', '2018 Practices', (SELECT id from csp_program_id),
+        FALSE),
+       ('Bundles', 'Bundles', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Soil Testing', 'Soil testing', '2018 Practices', (SELECT id from csp_program_id), False),
+       ('Other Improvement', 'Other improvement', '2018 Practices', (SELECT id from csp_program_id), FALSE),
+       ('Miscellaneous', 'Miscellaneous', 'Miscellaneous Practices', (SELECT id from csp_program_id), FALSE);
 
 -- Supplemental Nutrition Assistance Program (SNAP)
 INSERT INTO pdl.programs(name, title_id)