+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Feature to parse and ingest Title-II EQIP and CSP
+  data. [#7](https://github.com/policy-design-lab/data-ingestion/issues/7)
+
+### Changed
+
+- Create schema now uses a new command-line
+  argument. [#27](https://github.com/policy-design-lab/data-ingestion/issues/27)
+
+#### Fixed
+
+- Issue with the CLI not using password provided as command-line
+  argument. [#25](https://github.com/policy-design-lab/data-ingestion/issues/25)
+
 ## [0.2.0] - 2024-07-22
 
 ### Added